Add files via upload
New Wheels Project
Introduction to SQL
Problem Statement
Business Context
A lot of people in the world share a common desire: to own a vehicle. A car or an
automobile is seen as an object that gives the freedom of mobility. Many now prefer
pre-owned vehicles because they come at an affordable cost, but at the same time, they
are also concerned about whether the after-sales service provided by the resale
vendors is as good as the care you may get from the actual manufacturers.
New-Wheels, a vehicle resale company, has launched an app with an end-to-end
service from listing the vehicle on the platform to shipping it to the customer's location.
This app also captures the overall after-sales feedback given by the customer.
Objective
New-Wheels sales have been dipping steadily in the past year, and due to the critical
customer feedback and ratings online, there has been a drop in new customers every
quarter, which is concerning to the business. The CEO of the company now wants a
quarterly report with all the key metrics sent to him so he can assess the health of the
business and make the necessary decisions.
As a data analyst, you see that there is an array of questions that are being asked at
the leadership level that need to be answered using data. Import the dump file that
contains various tables that are present in the database. Use the data to answer the
questions posed and create a quarterly business report for the CEO.